feat: Add welcome message for new guild members
In the `welcomeMessage.js` file, a new event handler named `guildMemberAdd` was implemented. When a new member joins the guild, a welcome message is created using an EmbedBuilder and sent to the system channel. If the system channel is not found, an error message is logged.

In the `index.js` file, the client is now initialized with the `GuildMembers` intent in addition to the `Guilds` intent.