feat(events): add thread lock on resolve functionality
This commit adds a new feature to the `threadUpdate` module which allows for thread locking and unlocking based on the addition or removal of a specific tag. If the tag named 'Resolved' is added to a thread in the 'help-desk' category, the thread will be locked and a corresponding message will be sent. Conversely, if the 'Resolved' tag is removed, the thread will be unlocked and a different message will be sent.