feat: Add slash-eval command for evaluating JavaScript code
This commit adds a new feature - a slash command called "eval" that allows users to evaluate JavaScript code. The command accepts a required "code" option, which represents the code to be evaluated. Additionally, there is an optional "async" option that determines whether the code is asynchronous. The code is executed using the vm module in a sandboxed environment. If the code executes successfully, the result is displayed in an embed, and if an error occurs, the error message is shown in a separate embed. The command is only accessible to the application owner and team members.